Threads (#16)
* Creating Logger

* Created logger

* Added MessageLogger class

* Improved getting message id

* Added logger to Client class

* Created method networkMonitoring

* Saved message request

* Added method showMessages

* Refactorred loggers

* Added unlock

* Refactored message monitoring

* Updating Client class

* Finished Client class

* Added request for message by id

* Fixed bugs in loggers

* Problem with logger

* Fixed logger

* Refactored and fixed some locks

* Thread problem solved

* Fixin logger problems

* Fixed blocking logger

* Fixed logger positions

* Removed blocking output by waiting input

* Refactored file naming

* Improved Readme

* Fixed blockin at logout

+++ src/sql.cpp
@@ -178,7 +178,7 @@ std::shared_ptr<Message>SQL::getMessage(const uint32_t lastId, const std::string
 		to = std::string(row[2]);
 	}
 
-	return std::make_shared<Message>(row[0], from, to, row[3], atoi(row[4]));
+	return std::make_shared<Message>(row[0], from, to, row[3], static_cast<uint32_t>(std::stoul((row[4]))));
 }
 
 SQL &SQL::operator=(SQL &&that)noexcept
